Test configuration before restarting
The legacy configuration format is difficult to test with the
template.validate attribute. This is the next best thing.
Right before restart, we run --config=check. If it fails, the play
stops. Avoids putting the service in a boot-loop.

Skipped during check-mode, because we want to always check the new
configuration, never the previous one.

Something to consider when engineering the new configuration format.